moved some tool api stuff to item api
