+++ felt-item-api/build.gradle
@@ -1,0 +1,2 @@
+archivesBaseName = "felt-item-api"
+version = getSubprojectVersion(project)
