+++ felt-item-api/src/main/resources/fabric.mod.json
@@ -1,0 +1,31 @@
+{
+  "schemaVersion": 1,
+  "id": "felt-item-api",
+  "version": "${version}",
+  "name": "Felt Item API",
+  "description": "A reimplementation of Fringe APIs on Fabric",
+  "authors": [
+    "FeltMC",
+    "PseudoDistant",
+    "Trinsdar",
+    "Unix-Supremacist"
+  ],
+  "contact": {
+    "homepage": "",
+    "sources": "https://github.com/FeltMC/Felt-API"
+  },
+  "license": "LGPL",
+  "icon": "assets/felt-tool-api/icon.png",
+
+  "environment": "*",
+  "mixins": [
+    "itemapi.mixins.json"
+  ],
+  "depends": {
+    "fabricloader": "*",
+    "minecraft": "*"
+  },
+  "recommends": {
+    "feltfabric": "*"
+  }
+}
