+++ felt-item-api/src/main/resources/itemapi.mixins.json
@@ -1,0 +1,12 @@
+{
+  "required": true,
+  "minVersion": "0.8",
+  "package": "io.github.feltmc.feltapi.mixin.item",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "PacketByteBufMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
