+++ gradle.properties
@@ -15,6 +15,7 @@ org.gradle.jvmargs=-Xmx1G
 	felt-ore-feature-api-v1-version=1.0.0
 	felt-splash-api-version=1.0.0
 	felt-tool-api-version=1.0.0
+	felt-item-api-version=1.0.0
 	felt-shared-lifecycle-events-version=1.0.0
 
 # Dependencies
