+++ settings.gradle
@@ -15,4 +15,5 @@ include("felt-api-base")
 include("felt-ore-feature-api-v1")
 include("felt-splash-api")
 include("felt-tool-api")
+include("felt-item-api")
 include("felt-shared-lifecycle-events")
